web/requirements.txt
ca-certificates==2020.12.5
certifi==2020.12.5
click==7.1.2
flask==1.1.2
itsdangerous==1.1.0
jinja2==2.11.3
ld_impl_linux-64==2.35.1
libffi==3.3
libgcc-ng==9.3.0
libgomp==9.3.0
libstdcxx-ng==9.3.0
markupsafe==1.1.1
ncurses==6.2
openssl==1.1.1k
pip==21.0.1
python==3.8.8
python_abi==3.8
readline==8.0
setuptools==49.6.0
sqlite==3.35.4
tk==8.6.10
werkzeug==1.0.1
wheel==0.36.2
xz==5.2.5
zlib==1.2.11
pandas==1.2.4
numpy==1.20.0
matplotlib==3.4.1
scikit-learn==0.24.1
scipy==1.6.2
sklearn-pandas==2.1.0